:recycle: [operator] Move proof types to separate file
